requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
requests==2.28.1
djangorestframework==3.12.4